adjusting code after review & additional testing
Also moved `SchemaTools` object from `Tools.scala` into a separate file - it's easier to
maintain, no real changes in the code.
